# NibblesSwagger.yaml
swagger: "2.0"
info:
  description: "Un Nibbles en el que hay que el comportamiento de cada serpiente está programado utilizando una llamada REST. En este documento se explica el método y el comportamiento esperado para participar. Cada serpiente tiene una cabeza que se mueve en una dirección 'Up', 'Right', 'Down', 'Left', para cada turno se le solicita al jugador que indique el próximo movimiento en base al estado del juego que es informado junto con el requerimiento."
  version: "1.0.0"
  title: "Lagash Nibbles"
  termsOfService: "http://lagash.com/terms/"
  contact:
    email: "diegog@lagash.com"
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "nibbles.lagash.com"
basePath: "/"
schemes:
- "http"
paths:
  /nextMove:
    post:
      tags:
      - "nextMove"
      summary: "Devuelve la nueva dirección a la que debe dirigirse la serpiente."
      description: "No es necesario que el componente mantenga (recuerde) el estado del juego, ya que se le es informado en cada llamada."
      operationId: "nextMove"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "La información del estado del juego."
        required: true
        schema:
          $ref: "#/definitions/NibblesStatus"
      responses:
        405:
          description: "Invalid input"
definitions:
  NibblesStatus:
    description: "La información del estado del juego."
    type: "object"
    required:
    - "snake"
    - "space"
    - "snakes"
    properties:
      snake:
        description: "La serpiente que debe ser controlada."
        $ref: "#/definitions/Snake"
      space:
        description: "El mapa del espacio en el que se encuentran todos los jugadores."
        $ref: "#/definitions/Space"
      snakes:
        type: "array"
        items:
          $ref: "#/definitions/Snake"
        description: "El estado de todas las serpientes del juego."
  Snake:
    description: "La información de una serpiente en el juego."
    type: "object"
    required:
    - "id"
    - "x"
    - "y"
    - "direction"
    - "length"
    - "ticks"
    - "trail"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "El identificador único de la serpiente."
        example: "1"
      x:
        type: "integer"
        format: "int64"
        description: "La coordenada x de la serpiente. Dentro de los márgenes del espacio de juego."
        example: "12"
      y:
        type: "integer"
        format: "int64"
        description: "La coordenada y de la serpiente. Dentro de los márgenes del espacio de juego."
        example: "23"
      direction:
        type: "string"
        description: "La dirección en la que está moviéndose la serpiente."
        enum:
        - "Up"
        - "Left"
        - "Down"
        - "Right"
      length:
        type: "integer"
        format: "int64"
        description: "El largo total de la serpiente."
        example: "7"
      ticks:
        type: "integer"
        format: "int64"
        description: "La cantidad de movidas que lleva la serpiente."
        example: "6"
      trail:
        type: "array"
        items:
          $ref: "#/definitions/Vector"
        description: "La información del trazo de cada serpiente, donde cada elemento de la lista es un cambio de dirección."
  Space:
    description: "El mapa del estado del juego."
    type: "object"
    required:
    - "topX"
    - "topY"
    - "map"
    properties:
      topX:
        type: "integer"
        format: "int64"
        description: "El tamaño máximo horizontal del espacio de juego."
      topY:
        type: "integer"
        format: "int64"
        description: "El tamaño máximo vertical del espacio de juego."
      map:
        type: "array"
        description: "Una matriz con el mapa del juego, la matriz se interpreta map[x][y], donde cada valor indica 0 (si la casilla se encuentra vacía), ó el número de la serpiente que se encuentra en dicho casillero."
        items:
          type: "array"
          items:
            type: "integer"
            format: "int64"
  Vector:
    description: "Información de posición y dirección de movimiento."
    type: "object"
    properties:
      x:
        type: "integer"
        format: "int64"
        description: "La coordenada x de la serpiente."
        example: "9"
      y:
        type: "integer"
        format: "int64"
        description: "La coordenada y de la serpiente."
        example: "3"
      direction:
        type: "string"
        description: "La dirección en la que está moviéndose."
        enum:
        - "Up"
        - "Left"
        - "Down"
        - "Right"
externalDocs:
  description: "Mas información sobre Laagsh Nibbles."
  url: "http://nibbles.lagash.com"
  